+++ src/main/java/com/example/onlinebookstore/OnlinebookstoreApplication.java
@@ -1,0 +1,13 @@
+package com.example.onlinebookstore;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OnlinebookstoreApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OnlinebookstoreApplication.class, args);
+	}
+
+}
