--- src/main/java/mate/academy/onlinebookstore/OnlinebookstoreApplication.java
@@ -1,16 +1,0 @@
-package mate.academy.onlinebookstore;
-
-import mate.academy.onlinebookstore.service.BookService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OnlinebookstoreApplication {
-    @Autowired
-    private BookService bookService;
-
-    public static void main(String[] args) {
-        SpringApplication.run(OnlinebookstoreApplication.class, args);
-    }
-}
