--- src/main/java/mate/academy/onlinebookstore/config/MapperConfig.java
@@ -1,14 +1,0 @@
-package mate.academy.onlinebookstore.config;
-
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.NullValueCheckStrategy;
-
-@org.mapstruct.MapperConfig(
-        componentModel = "spring",
-        injectionStrategy = InjectionStrategy.CONSTRUCTOR,
-        nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS,
-        implementationPackage = "<PACKAGE_NAME>.impl"
-)
-public class MapperConfig {
-
-}
