--- src/main/java/mate/academy/onlinebookstore/controller/BookController.java
@@ -1,36 +1,0 @@
-package mate.academy.onlinebookstore.controller;
-
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import mate.academy.onlinebookstore.dto.BookDto;
-import mate.academy.onlinebookstore.dto.CreateBookRequestDto;
-import mate.academy.onlinebookstore.service.BookService;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping(value = "/books")
-public class BookController {
-
-    private final BookService bookService;
-
-    @GetMapping
-    public List<BookDto> getAll() {
-        return bookService.findAll();
-    }
-
-    @GetMapping("/{id}")
-    public BookDto getBookById(@PathVariable Long id) {
-        return bookService.findById(id);
-    }
-
-    @PostMapping
-    public BookDto createBook(@RequestBody CreateBookRequestDto bookDto) {
-        return bookService.save(bookDto);
-    }
-}
