--- src/main/java/mate/academy/onlinebookstore/dto/BookDto.java
@@ -1,14 +1,0 @@
-package mate.academy.onlinebookstore.dto;
-
-import lombok.Data;
-
-@Data
-public class BookDto {
-    private Long id;
-    private String title;
-    private String author;
-    private String isbn;
-    private Double price;
-    private String description;
-    private String coverImage;
-}
