--- src/main/java/mate/academy/onlinebookstore/dto/CreateBookRequestDto.java
@@ -1,13 +1,0 @@
-package mate.academy.onlinebookstore.dto;
-
-import lombok.Data;
-
-@Data
-public class CreateBookRequestDto {
-    private String title;
-    private String author;
-    private String isbn;
-    private Double price;
-    private String description;
-    private String coverImage;
-}
