--- src/main/java/mate/academy/onlinebookstore/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package mate.academy.onlinebookstore.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
