--- src/main/java/mate/academy/onlinebookstore/mapper/BookMapper.java
@@ -1,14 +1,0 @@
-package mate.academy.onlinebookstore.mapper;
-
-import mate.academy.onlinebookstore.config.MapperConfig;
-import mate.academy.onlinebookstore.dto.BookDto;
-import mate.academy.onlinebookstore.dto.CreateBookRequestDto;
-import mate.academy.onlinebookstore.entity.Book;
-import org.mapstruct.Mapper;
-
-@Mapper(config = MapperConfig.class)
-public interface BookMapper {
-    BookDto toDto(Book book);
-
-    Book toModel(CreateBookRequestDto requestDto);
-}
