--- src/main/java/mate/academy/onlinebookstore/repository/BookRepository.java
@@ -1,13 +1,0 @@
-package mate.academy.onlinebookstore.repository;
-
-import java.util.List;
-import java.util.Optional;
-import mate.academy.onlinebookstore.entity.Book;
-
-public interface BookRepository {
-    Book save(Book book);
-
-    Optional<Book> findById(Long id);
-
-    List<Book> findAll();
-}
