--- src/main/java/mate/academy/onlinebookstore/repository/impl/BookRepositoryImpl.java
@@ -1,51 +1,0 @@
-package mate.academy.onlinebookstore.repository.impl;
-
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.EntityTransaction;
-import java.util.List;
-import java.util.Optional;
-import lombok.RequiredArgsConstructor;
-import mate.academy.onlinebookstore.entity.Book;
-import mate.academy.onlinebookstore.repository.BookRepository;
-import org.springframework.stereotype.Repository;
-
-@RequiredArgsConstructor
-@Repository
-public class BookRepositoryImpl implements BookRepository {
-
-    private final EntityManagerFactory entityManagerFactory;
-
-    @Override
-    public Book save(Book book) {
-        EntityTransaction transaction = null;
-        try (EntityManager entityManager = entityManagerFactory.createEntityManager()) {
-            transaction = entityManager.getTransaction();
-            transaction.begin();
-            entityManager.persist(book);
-            transaction.commit();
-            return book;
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            throw new RuntimeException("Can't insert a book : " + book, e);
-        }
-    }
-
-    @Override
-    public Optional<Book> findById(Long id) {
-        try (EntityManager entityManager = entityManagerFactory.createEntityManager()) {
-            Book book = entityManager.find(Book.class, id);
-            return book != null ? Optional.of(book) : Optional.empty();
-        }
-    }
-
-    @Override
-    public List<Book> findAll() {
-        try (EntityManager entityManager = entityManagerFactory.createEntityManager()) {
-            return entityManager.createQuery("Select b from Book b",
-                    Book.class).getResultList();
-        }
-    }
-}
