--- src/main/java/mate/academy/onlinebookstore/service/BookService.java
@@ -1,14 +1,0 @@
-package mate.academy.onlinebookstore.service;
-
-import java.util.List;
-import mate.academy.onlinebookstore.dto.BookDto;
-import mate.academy.onlinebookstore.dto.CreateBookRequestDto;
-
-public interface BookService {
-
-    BookDto save(CreateBookRequestDto requestDto);
-
-    BookDto findById(Long id);
-
-    List<BookDto> findAll();
-}
