--- src/main/java/mate/academy/onlinebookstore/service/impl/BookServiceImpl.java
@@ -1,42 +1,0 @@
-package mate.academy.onlinebookstore.service.impl;
-
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import mate.academy.onlinebookstore.dto.BookDto;
-import mate.academy.onlinebookstore.dto.CreateBookRequestDto;
-import mate.academy.onlinebookstore.entity.Book;
-import mate.academy.onlinebookstore.exception.EntityNotFoundException;
-import mate.academy.onlinebookstore.mapper.BookMapper;
-import mate.academy.onlinebookstore.repository.BookRepository;
-import mate.academy.onlinebookstore.service.BookService;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class BookServiceImpl implements BookService {
-
-    private final BookRepository bookRepository;
-
-    private final BookMapper bookMapper;
-
-    @Override
-    public BookDto save(CreateBookRequestDto requestDto) {
-        Book book = bookMapper.toModel(requestDto);
-        return bookMapper.toDto(bookRepository.save(book));
-    }
-
-    @Override
-    public BookDto findById(Long id) {
-        Book book = bookRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Can't find book by id" + id)
-        );
-        return bookMapper.toDto(book);
-    }
-
-    @Override
-    public List<BookDto> findAll() {
-        return bookRepository.findAll().stream()
-                .map(bookMapper::toDto)
-                .toList();
-    }
-}
