+++ src/main/resources/application.properties
@@ -1,7 +1,1 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/ticket_app?serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=root
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-hibernate.dialect=org.hibernate.dialect.MySQLDialect
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.show-sql=true
+
