+++ src/test/java/com/example/onlinebookstore/OnlinebookstoreApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.onlinebookstore;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OnlinebookstoreApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
