--- src/test/java/mate/academy/onlinebookstore/OnlinebookstoreApplicationTests.java
@@ -1,11 +1,0 @@
-package mate.academy.onlinebookstore;
-
-import org.junit.jupiter.api.Test;
-
-class OnlinebookstoreApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
